roman numerals - finished for now
I've avoided this problem for ages, it has no interest
for me in any way. But I needed to get through it as
part of the progression, so here it is.

It could be made a lot simpler, but I just don't have the
interest in it. Perhaps some other time.