nth prime, by brute force
There is no good way to determine the nth prime, if
there were, there would not be some much research
activity around prime numbers. There are ways to
estimate some numbers very close to the nth prime,
which I believe might be overkill for this problem.

I would expect the if there was intent to help the
programming understand more about prime numbers,
the README would be a bit more informative.